+++ CHANGELOG.md
@@ -11,7 +11,9 @@ The [releases page](https://github.com/Jelmerro/Vieb/releases) also contains the
 but the list below contains much more technical details.
 The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
-## Unreleased
+## [10.4.0](https://github.com/Jelmerro/Vieb/compare/10.3.0...10.4.0) - 2023-10-12
+
+[code diff](https://github.com/Jelmerro/Vieb/compare/10.3.0...10.4.0) - [released builds](https://github.com/Jelmerro/Vieb/releases/tag/10.4.0)
 
 ### Added